refactor: rename base to compressor_consumer_function
