+++ src/libecalc/domain/process/compressor/dto/__init__.py
@@ -1,5 +1,5 @@
-from .base import CompressorConsumerFunction
 from .chart import CompressorChart
+from .compressor_consumer_function import CompressorConsumerFunction
 from .sampled import CompressorSampled
 from .stage import CompressorStage, InterstagePressureControl, MultipleStreamsCompressorStage
 from .train import (
