+++ src/libecalc/domain/process/compressor/dto/compressor_consumer_function.py
